Conflicts:
	ElevatorSim.cpp
	elevatorSim.sln

+++ ElevatorSim.hpp
@@ -51,29 +51,29 @@
 #include <cmath>
 
 const float M_PI = 3.141592f;
-const double FPS = (double)1/(double)60;
+const double FPS = 1/(double)60;
 
 struct Vec3f {
-	float x, y, z;
+   float x, y, z;
 };
 
 enum KEY_PRESS {
-	KEY_UP,
-	KEY_DOWN,
-	KEY_RIGHT,
-	KEY_LEFT,
-	KEY_FORWARD,
-	KEY_BACKWARD,
+   KEY_UP,
+   KEY_DOWN,
+   KEY_RIGHT,
+   KEY_LEFT,
+   KEY_FORWARD,
+   KEY_BACKWARD,
 };
 
 enum OBJ_LIST {
-	OBJ_CUBE = 1,
+   OBJ_CUBE = 1,
 };
 
 const GLfloat light1_ambient[]	= { 0.4f, 0.4f, 0.4f, 1.0f };
 const GLfloat light1_diffuse[]	= { 0.8f, 0.8f, 0.8f, 1.0f };
 const GLfloat light1_specular[]	= { 0.5f, 0.5f, 0.5f, 1.0f };
 const GLfloat light1_position[]	= { 3.0f, 10.0f, 3.0f, 0.0f };
-const GLfloat light1_direction[]	= { 0.0f, 0.0f, 0.0f, 0.0f };
+const GLfloat light1_direction[] = { 0.0f, 0.0f, 0.0f, 0.0f };
 
 #endif
